moved de-duped fetch-tx-proof logic
